feat: Create a branching tube demonstration
This change updates the `demo_branching_sdf.py` script to create a visualization of a branching tube structure.

The main changes are:
- The script now uses the `tube_sdf` function to create tubes for the trunk and branch curves.
- The SDFs of the individual tubes are combined using `union_sdf` to create a single, seamless branching structure.
- I removed the local `plot` function from the script, and it now uses the `plot_curve_sdf` function from the `sdf_utils` package to eliminate duplicate code.
